.streamlit/requirements.txt
streamlit
openai==0.28.0
pdfplumber
numpy
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-docx
python-dotenv
python-pptx
openpyxl
pandas
pillow
pydub
ffmpeg
speechrecognition
anthropic
 pypdf
 PdfReader
sentence-transformers==2.6.1
torch            # a versão CPU já é instalada como dependência
